+++ src/aStar.hpp
@@ -29,10 +29,10 @@ struct aStrCoordinate {
 };
 
 // Define a class for a node
-std::vector<aStrCoordinate> astar_path(aStrCoordinate&, aStrCoordinate&, short map[MAP_WIDTH][MAP_HEIGHT]);
+std::vector<aStrCoordinate> astar_path(aStrCoordinate&, aStrCoordinate&, uint8_t map[MAP_WIDTH][MAP_HEIGHT]);
 
 bool contains(std::vector<aStrCoordinate>&, aStrCoordinate&);
 
-std::vector<aStrCoordinate> getAdjacent(aStrCoordinate&, aStrCoordinate&, short map[MAP_WIDTH][MAP_HEIGHT]);
+std::vector<aStrCoordinate> getAdjacent(aStrCoordinate&, aStrCoordinate&, uint8_t map[MAP_WIDTH][MAP_HEIGHT]);
 
 float heuristic(int, int, int, int);
